Moved out jsx from MemberBookingsPage and made them into smaller components
